@@ -10,7 +10,7 @@
     }
   ],
   "require": {
-    "php": ">=7.3",
+    "php": ">=7.4",
     "ext-mbstring": "*",
     "ext-json": "*",
     "tightenco/collect": "^6.4"
